Moved the uncertainty estimation tests from test_integration.py to new file test_uncertainty.py.
